Added CAD of optics around the cryostat
The CAD of the optics around the Pasqal cryostat are added. This includes the optics for the absorption/transmission/reference spectroscopy setup as well as the optics for the COLBERT experiment. Two main setups are included. The first is the setup in the COLBERT transmission configuration while the second is the setup in the COLBERT reflection configuration.